Initial extraction of relationship queries.
