+++ src/Illuminate/Database/Query/Builder.php
@@ -1861,7 +1861,7 @@ class Builder
             // On each chunk result set, we will pass them to the callback and then let the
             // developer take care of everything within the callback, which allows us to
             // keep the memory low for spinning through large result sets for working.
-            if (call_user_func($callback, $results) === false) {
+            if ($callback($results) === false) {
                 return false;
             }
 
@@ -1903,7 +1903,7 @@ class Builder
             // On each chunk result set, we will pass them to the callback and then let the
             // developer take care of everything within the callback, which allows us to
             // keep the memory low for spinning through large result sets for working.
-            if (call_user_func($callback, $results) === false) {
+            if ($callback($results) === false) {
                 return false;
             }
